some minor cleanup ig

+++ zombies/src/main/java/org/phantazm/zombies/map/action/round/SelectBombedRoom.java
@@ -205,7 +205,7 @@ public class SelectBombedRoom implements Action<Round> {
                     }
                 }
             }
-        }, (long)data.gracePeriod);
+        }, data.gracePeriod);
     }
 
     private void applyModifiers(ZombiesPlayer player) {
